Remove ConnClient, it's pointless now (#1453)
